requirements.txt
gTTS==1.2.2
Flask==0.12.2
pychromecast==0.8.2
unicode_slugify==0.1.3
pathlib==1.0.1
samsungctl
mysqlclient
python-crontab
rpi_rf
